testing task v 1.0

+++ my_service/services/my_service.h
@@ -31,5 +31,5 @@ struct my_service_cb
 };
 
 int my_service_init(void);
-
-void my_service_send(struct bt_conn *conn, const uint8_t *data, uint16_t len);
+//void my_service_send(struct bt_conn *conn, const uint8_t *data, uint16_t len);
+void my_service_send(struct bt_conn *conn, const uint16_t *data, uint16_t len);
